+++ test/unit/parser/test_parser_function.py
@@ -19,16 +19,32 @@ def content():
         "    console.log('something_else');\n"
         "};\n"
         "\n"
-        "export default const doSomethingWithManyArguments = ("
-        "  arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9"
+        "export const doSomethingWithManyArguments = (\n"
+        "  arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9\n"
         ") => {};\n"
         "\n"
-        ""
+        "/**\n"
+        " * test anonymous function.\n"
+        " */\n"
+        "export default function (arg1) {\n"
+        "    console.log('anonymous function');\n"
+        "};\n"
+        "\n"
+        "/**\n"
+        " * test generator function.\n"
+        " */\n"
+        "function* generate(arg1, arg2) {\n"
+        "    yield something();\n"
+        "};\n"
     )
 
 
 def test_get_function_environment(content):
     """Return function environment from content."""
+    import pprint
+    pprint.pprint(champollion.parser.function_parser.get_function_environment(
+        content, "test.module"
+    ))
     assert champollion.parser.function_parser.get_function_environment(
         content, "test.module"
     ) == {
@@ -38,6 +54,8 @@ def test_get_function_environment(content):
             "exported": True,
             "default": False,
             "name": "doSomething",
+            "anonymous": False,
+            "generator": False,
             "arguments": [],
             "line_number": 2,
             "description": "test function"
@@ -48,6 +66,8 @@ def test_get_function_environment(content):
             "exported": False,
             "default": False,
             "name": "doSomethingElse",
+            "anonymous": False,
+            "generator": False,
             "arguments": ["arg1", "arg2"],
             "line_number": 9,
             "description": "test function with arguments."
@@ -56,8 +76,10 @@ def test_get_function_environment(content):
             "id": "test.module.doSomethingWithManyArguments",
             "module_id": "test.module",
             "exported": True,
-            "default": True,
+            "default": False,
             "name": "doSomethingWithManyArguments",
+            "anonymous": False,
+            "generator": False,
             "arguments": [
                 "arg1", "arg2", "arg3", "arg4",
                 "arg5", "arg6", "arg7", "arg8",
@@ -65,5 +87,29 @@ def test_get_function_environment(content):
             ],
             "line_number": 13,
             "description": None
+        },
+        "test.module.__ANONYMOUS_FUNCTION__": {
+            "id": "test.module.__ANONYMOUS_FUNCTION__",
+            "module_id": "test.module",
+            "exported": True,
+            "default": True,
+            "name": "__ANONYMOUS_FUNCTION__",
+            "anonymous": True,
+            "generator": False,
+            "arguments": ["arg1"],
+            "line_number": 20,
+            "description": "test anonymous function."
+        },
+        "test.module.generate": {
+            "id": "test.module.generate",
+            "module_id": "test.module",
+            "exported": False,
+            "default": False,
+            "name": "generate",
+            "anonymous": False,
+            "generator": True,
+            "arguments": ["arg1", "arg2"],
+            "line_number": 27,
+            "description": "test generator function."
         },
     }